XD-1011 add OModelMetaDataTest
